requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
diskcache==5.6.3
Flask==3.0.0
frozenlist==1.4.0
gptcache==0.1.42
guidance==0.0.64
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.24.1
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.1
platformdirs==3.11.0
pycparser==2.21
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
tiktoken==0.5.1
tqdm==4.66.1
urllib3==2.0.6
Werkzeug==3.0.0
yarl==1.9.2
